api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
slowapi==0.1.7
pydantic==2.3.0
python-dotenv==1.0.0 